"play:dependencies" mojo - added "dependenciesSkipJars" parameter. If "false" (default), jar type dependencies are copied to module's "lib" or "modules/${module}/lib" directories. If "true" - they are not copied. They are necessary for Play! Framework, but not needed for Maven build (Maven build uses only dependency mechanism).
git-svn-id: http://maven-play-plugin.googlecode.com/svn/trunk/plugin@5137 9c2ad861-7886-cb7c-dcd7-2b8d41cd5de0

+++ play-maven-plugin/src/main/java/com/google/code/play/PlayDependenciesMojo.java
@@ -68,6 +68,15 @@ public class PlayDependenciesMojo
     private boolean dependenciesOverwrite = false;
 
     /**
+     * Should jar dependencies be processed. They are necessary for Play! Framework,
+     * but not needed for Maven build (Maven uses dependency mechanism).
+     * 
+     * @parameter expression="${play.dependenciesSkipJars}" default-value="false"
+     * @since 1.0.0
+     */
+    private boolean dependenciesSkipJars = false;
+
+    /**
      * To look up Archiver/UnArchiver implementations.
      * 
      * @component role="org.codehaus.plexus.archiver.manager.ArchiverManager"
@@ -83,12 +92,18 @@ public class PlayDependenciesMojo
             if ( dependenciesClean )
             {
                 File baseDir = project.getBasedir();
-                FileUtils.deleteDirectory( new File( baseDir, "lib" ) );
+                if ( !dependenciesSkipJars )
+                {
+                    FileUtils.deleteDirectory( new File( baseDir, "lib" ) );
+                }
                 FileUtils.deleteDirectory( new File( baseDir, "modules" ) );
             }
 
             Map<Artifact, File> moduleTypeArtifacts = processModuleDependencies();
-            processJarDependencies( moduleTypeArtifacts );
+            if ( !dependenciesSkipJars )
+            {
+                processJarDependencies( moduleTypeArtifacts );
+            }
         }
         catch ( ArchiverException e )
         {
